+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ADD: Barcode generation in batch file
 - ADD: Show CAS number in formula
 - ADD: Dilutant field in finished product page
+- ADD: Include allergen ingredients in batch pdf file
 - FIX: Not showing allergens correctly in label
 - FIX: Message position in Generate Finished Product page
 - FIX: Various bug fixes